# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Documentation"
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get Events list",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer Token"
          },
          {
            "name": "page",
            "in": "query",
            "type": "number",
            "default": 1
          },
          {
            "name": "limit",
            "in": "query",
            "type": "number",
            "default": 10
          },
          {
            "name": "search",
            "in": "query",
            "type": "string",
            "description": "Search by event name and details info"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Fail"
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Create event",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "name": "language",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer Token"
          },
          {
            "name": "start_date",
            "in": "formData",
            "type": "string",
            "format": "date",
            "required": true
          },
          {
            "name": "end_date",
            "in": "formData",
            "type": "string",
            "format": "date",
            "required": true
          },
          {
            "name": "longitude",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "latitude",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "name",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "details",
            "in": "formData",
            "type": "string",
            "required": true
          },
          {
            "name": "category_id",
            "in": "formData",
            "type": "number",
            "required": true
          },
          {
            "name": "speaker_id",
            "in": "formData",
            "type": "number",
            "required": true
          },
          {
            "name": "cover_image",
            "in": "formData",
            "type": "file",
            "required": true
          },
          {
            "name": "images[]",
            "in": "formData",
            "type": "file",
            "multipleOf": 10,
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Fail"
          }
        }
      }
    },
    "/api/events/{id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get event by id",
        "parameters": [
          {
            "name": "language",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer Token"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Fail"
          }
        }
      },
      "put": {
        "tags": [
          "Events"
        ],
        "summary": "Update event",
        "parameters": [
          {
            "name": "language",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer Token"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "start_date",
            "in": "formData",
            "type": "string",
            "format": "date"
          },
          {
            "name": "end_date",
            "in": "formData",
            "type": "string",
            "format": "date"
          },
          {
            "name": "longitude",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "latitude",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "details",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "category_id",
            "in": "formData",
            "type": "number"
          },
          {
            "name": "speaker_id",
            "in": "formData",
            "type": "number"
          },
          {
            "name": "cover_image",
            "in": "formData",
            "type": "file"
          },
          {
            "name": "images[]",
            "in": "formData",
            "type": "file",
            "multipleOf": 10
          },
          {
            "name": "delete_image_ids",
            "in": "formData",
            "type": "string",
            "description": "Example` 1,2,3,18"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Fail"
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "summary": "Delete event",
        "parameters": [
          {
            "name": "language",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Bearer Token"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Fail"
          }
        }
      }
    },
...
  }
}